+++ client/src/game/api/emits/players.ts
@@ -1,6 +1,6 @@
-import type { PlayerRoleSet, PlayersBring } from "../../../apiTypes";
+import type { PlayerRoleSet, PlayerPosition } from "../../../apiTypes";
 import { wrapSocket } from "../helpers";
 
-export const sendBringPlayers = wrapSocket<PlayersBring>("Players.Bring");
+export const sendBringPlayers = wrapSocket<PlayerPosition>("Players.Bring");
 
 export const sendChangePlayerRole = wrapSocket<PlayerRoleSet>("Player.Role.Set");