+++ server/src/api/socket/player.py
@@ -9,14 +9,14 @@ from ...logs import logger
 from ...models.role import Role
 from ...state.game import game_state
 from ..helpers import _send_game
-from ..models.players import PlayersBring
+from ..models.players import PlayerPosition
 from ..models.players.role import PlayerRoleSet
 
 
 @sio.on("Players.Bring", namespace=GAME_NS)
 @auth.login_required(app, sio, "game")
 async def bring_players(sid: str, raw_data: Any):
-    data = PlayersBring(**raw_data)
+    data = PlayerPosition(**raw_data)
 
     pr = game_state.get(sid)